Add: Centralize types
