+++ backend/src/models/User.ts
@@ -5,10 +5,7 @@ import { IUser } from '../../../shared/user.interface'; ;
 
 // Define IUserDocument as a HydratedDocument of IUser
 // This correctly gives it all properties of IUser AND Mongoose Document methods/properties. HydratedDocument is a utility type. It takes the IUser and returns a new type including Mongoose's Types with the one of IUser so the ID from Mongoose is used instead of the string ID from the interface.
-export interface IUserDocument extends HydratedDocument<IUser> {
-    passwordHash: string; 
-    matchPassword(enteredPassword: string): Promise<boolean>; 
-}
+import { IUserDocument } from '../types/declarations';
 const UserSchema: Schema<IUserDocument> = new Schema({
     username: {
         type: String,
